crates/altius_py/pyproject.toml
[project]
name = "altius_py"
version = "0.1.0"
authors = [
    { name = "maekawatoshiki" }
]
dependencies = [
    "pillow>=9.2.0",
    "matplotlib>=3.5.3",
    "maturin>=1.1.0",
    "onnx>=1.15.0",
    "transformers==4.41.2",
    "onnxsim==0.4.17",
    "numpy>=1.25.2",
    "onnxruntime>=1.15.1",
    "pytest>=7.4.4",
    "pytest-xdist>=3.3.1",
    "torch>=2.2.2",
    "onnxscript>=0.1.0.dev20240227",
    "pip>=24.0",
    "optimum>=1.18.1",
    "torchvision>=0.17.2",
    "tabulate>=0.9.0",
    "timm>=0.9.16",
    "packaging>=24.1",
    "einops>=0.8.0",
    "fpzip>=1.2.4",
    "zfpy>=1.0.0",
    "onnxruntime-genai>=0.3.0; sys_platform == 'linux'",
    "huggingface>=0.0.1",
    "netron>=7.8.3",
    "patchelf>=0.17.2.1; sys_platform == 'linux'",
]
requires-python = ">=3.12"

[build-system]
requires = ["maturin>=1.1.0"]
build-backend = "maturin"
